tuJanalyst/pyproject.toml
[project]
name = "tujanalyst"
version = "0.1.0"
description = "AI-powered stock analysis system for NSE/BSE listed companies"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Database
    "motor>=3.6.0",                  # Async MongoDB driver
    "pymongo>=4.9.0",               # MongoDB driver (motor dependency)

    # Vector store
    "chromadb>=0.5.0",
    "sentence-transformers>=3.2.0",

    # LLM & AI
    "dspy-ai>=2.5.0",
    "pydantic-ai>=0.0.30",
    "anthropic>=0.40.0",

    # Data processing
    "pdfplumber>=0.11.0",           # PDF text extraction
    "beautifulsoup4>=4.12.0",       # HTML parsing
    "lxml>=5.0.0",                  # Fast XML/HTML parser
    "yfinance>=0.2.50",             # Market data

    # HTTP & feeds
    "httpx>=0.27.0",                # Async HTTP client
    "feedparser>=6.0.0",            # RSS feed parsing

    # Scheduling
    "apscheduler>=3.10.0",          # Background job scheduling

    # Configuration
    "pydantic-settings>=2.6.0",     # Settings from env vars
    "pyyaml>=6.0.0",               # YAML config loading

    # Utilities
    "python-multipart>=0.0.12",     # FastAPI form data
    "python-dotenv>=1.0.0",         # .env file loading
    "structlog>=24.4.0",            # Structured JSON logging
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "mongomock-motor>=0.0.30",      # Mock MongoDB for tests
    "ruff>=0.8.0",                  # Linting & formatting
    "mypy>=1.13.0",                 # Type checking
]


[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
